Revert "ssh-key: move `Signature` into `certificate` module (#24)"
This reverts commit 24fd0fe79719fe04c9bdc5869b4ac6cf611169e1.

It turns out the `Signature` format used by `sshsig` is the same shape
as the one used by certificates.

This means the encoding logic as well as the signing/verification logic
can all be centralized in the `Signature` type, with `SshSig` as
higher-level wrapper type composed in terms of `Signature`.