ssh-encoding: unify PEM encoding in `EncodePem` trait
This commit adds a "heapless" mode for `EncodePem` which writes the
encoded output to a provided buffer.

The old `EncodePem::encode_pem` method is renamed to `encode_pem_string`
which is composed in terms of the new `::encode_pem` method.

This replaces the previous logic for doing this in the `ssh-key` crate.