ssh-encoding: remove `Encoding::Error`; add `LabelError`
Removes the associated `Encoding::Error` type as well as `Label::Error`,
using `ssh_encoding::Error` as the type.

Adds a `LabelError` type which captures the invalid label when the
`alloc` feature has been enabled for display purposes. This is used as
the error type for `FromStr` impls, so invalid labels can be propagated
properly.