ssh-key: `p521` feature
Adds initial integration with the `p521` crate, including support for
converting public and private keys to `p521` keys, as well as signing
and verifying NIST P-521 signatures.

Note that there's a lot of redundancy/boilerplate repeated for `p256`,
`p384`, and `p521` it would be nice to eventually DRY out, possibly
using macros to write the impls.