Add newline to `PublicKey::write_openssh_file` output
Also adds an integration test to ensure that the file output matches
what `ssh-keygen` generates byte-for-byte.

The `ToString` impl still omits the newline.

To avoid making breaking changes, it always uses `\n` as the newline,
however perhaps in a future breaking release it could be changed to
accept `LineEnding` so as to match `PrivateKey::write_openssh_file`.