ssh-encoding: remove `NestedReader`
Changes `Reader::read_prefixed` from a provided method to one which must
be impl'd by all of the `Reader` types, of which there are currently
three:

- `&[u8]`
- `Base64Reader`
- `PemReader`

This should reduce monomorphization bloat from `NestedReader` which
encapsulates other reader types, including other `NestedReader`s,
depending on the level of nesting, which is generally low, but good to
eliminate entirely.

The `Reader` impl is written identically between `Base64Reader` and
`PemReader` and currently impl'd as a macro to prevent duplication,
however a more elegant solution would be nice, e.g. a generic method.