ssh-cipher: make ChaCha20Poly1305 accept a 256-bit key
In the OpenSSH protocol, ChaCha20Poly1305 uses two different keys: one
for length encryption, and one for payloads.

It wasn't previously possible to use both keys (or at least quite
obtuse) because `ssh-cipher` chopped up `k_1` and `k_2` and only allowed
`k_1` to be used.

This change hoists that into `ssh-key` for file encryption uses, giving
access to the raw `ChaCha20Poly1305` implementation and allowing use of
either key.

In addition, a nonce is now mandatory, but use of the default all-zero
nonce for file encryption purposes is now handled in `ssh-key`.