ssh-cipher: allows to decrypt a payload without altering state (#266)
When decrypting messages from the wire, we need to first decrypt the
first couple of bytes (4 or 16 bytes) just to get the length of packet.

Once we get the length of the packet we can decrypt the full payload. If
we just used the Decryptor api as it is, we would alter the internal
state.