ssh-encoding: fix `string` comments on encoding trait impls
The impls of `Decode` for `Bytes` and `Encode` for `[u8]` made reference
to `byte[n]`, i.e. a fixed length bytestring, however both these types
are analogues for `string`, which is encoded with a length prefix.

The comments have been updated to reflect the correct type they map to
in the SSH encoding schema, which is `string`.