ssh-key: extract a `Comment` type
Extracts a type which models the concerns of comments, namely that they
can be encoded as an RFC4251 `string` type which can hold arbitrary
binary data when stored in the binary serialization of a private key.

This replaces the various accessor methods on `PublicKey` and
`PrivateKey` for handling the various conversions.