Implement Encode for sequences of string-like types
With this change, various types like `&[&str]`, `Vec<&String>,
&[Vec<u8>`, etc. are now `Encode` by default, and encode as an RFC4251
string of the encoded entries.

Custom types can opt into this by implementing the marker trait `Rfc4251String`.

Implementation note: This would be more general and cover more types if
we could blanket `impl<T: Encode> Encode for &T`, as this would cover
any level of references in the trait bound for the `Rfc4251String`
blanket implenentation. However, this would collide with the `Label`
trait, so instead this adds explicit impls for the immediate types that
we implement `Encode` for.