ssh-key: use `std::env::home_dir` instead of `home` crate (#415)
Prior to Rust 1.86, `std::env::home_dir` was deprecated in favor of the
`home` crate. However, Rust 1.86 fixed the behavior of the function on
Windows to use the `USERPROFILE` environment variable with a fallback to
`GetUserProfileDirectory`, meaning we no longer need to use a crate for
this purpose.

Also, `home` now has an MSRV of 1.88, which would complicate our plans
to launch with an MSRV of 1.85.